Purchases: Rename helpers to utils to be consistent with other places in Calypso
